moved 'copy java problems' configurable issue code to xtext core https://github.com/eclipse/xtext-core/issues/544
