+++ org.eclipse.xtext/src/org/eclipse/xtext/validation/IssueCodes.java
@@ -1,0 +1,22 @@
+/*******************************************************************************
+ * Copyright (c) 2017 itemis AG (http://www.itemis.eu) and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *******************************************************************************/
+package org.eclipse.xtext.validation;
+
+/**
+ * @author Christian Dietrich - Initial contribution and API
+ * @since 2.14
+ */
+public class IssueCodes {
+	
+	/**
+	 * IssueCode to control DerivedResourceMarkerCopier
+	 */
+	// @see org.eclipse.xtext.builder.smap.DerivedResourceMarkerCopier
+	public static final String COPY_JAVA_PROBLEMS = "org.eclipse.xtext.builder.copyJavaProblems";
+
+}
